Step on Shawns toes, started showing graphs
