+++ wandb/__init__.py
@@ -298,6 +298,7 @@ def init(job_type='train', config=None):
         termlog(
             'wandb dry run mode. Run `wandb board` from this directory to see results')
         termlog()
+        run.config.set_run_dir(run.dir)
         _init_headless(api, run, job_type, False)
     else:
         termlog(
